Rename `Catalog` to `Pool`.
The "pool" name feels more familiar.
