+++ cap-async-std/src/lib.rs
@@ -6,22 +6,22 @@
 //! objects which can be passed around between different parts of a
 //! program.
 //!
-//! Two notable features are the [`Dir`] and [`Catalog`] types:
+//! Two notable features are the [`Dir`] and [`Pool`] types:
 //!  - `Dir` represents an open directory in a filesystem. Instead of
 //!    opening files by absolute paths or paths relative to the current
 //!    working directory, files are opened via paths relative to a
 //!    `Dir`. The concepts of a process-wide "current working directory"
 //!    and a single global filesystem namespace are de-emphasized.
-//!  - `Catalog` represents a set of network addresses. Instead of
+//!  - `Pool` represents a set of network addresses. Instead of
 //!    allowing applications to request access to any address and then
 //!    applying process-wide filtering rules, filtering rules are
-//!    built into catalogs which may be passed through the program.
+//!    built into pools which may be passed through the program.
 //!
 //! On WASI, use of this library closely reflects the underlying system
 //! API, so it avoids compatibility layers.
 //!
 //! [`Dir`]: fs::Dir
-//! [`Catalog`]: net::Catalog
+//! [`Pool`]: net::Pool
 
 #![deny(missing_docs)]
 #![cfg_attr(target_os = "wasi", feature(wasi_ext))]
