+++ cap-async-std/src/net/tcp_stream.rs
@@ -19,11 +19,11 @@ use {
 /// This corresponds to [`async_std::net::TcpStream`].
 ///
 /// Note that this `TcpStream` has no `connect` method. To create a `TcpStream`,
-/// you must first obtain a [`Catalog`] permitting the address, and then call
-/// [`Catalog::connect_tcp_stream`].
+/// you must first obtain a [`Pool`] permitting the address, and then call
+/// [`Pool::connect_tcp_stream`].
 ///
-/// [`Catalog`]: struct.Catalog.html
-/// [`Catalog::connect_tcp_stream`]: struct.Catalog.html#method.connect_tcp_stream
+/// [`Pool`]: struct.Pool.html
+/// [`Pool::connect_tcp_stream`]: struct.Pool.html#method.connect_tcp_stream
 pub struct TcpStream {
     std: net::TcpStream,
 }
