+++ cap-std/src/net/tcp_listener.rs
@@ -14,11 +14,11 @@ use {
 /// This corresponds to [`std::net::TcpListener`].
 ///
 /// Note that this `TcpListener` has no `bind` method. To bind it to a socket
-/// address, you must first obtain a [`Catalog`] permitting the address, and
-/// then call [`Catalog::bind_tcp_listener`].
+/// address, you must first obtain a [`Pool`] permitting the address, and
+/// then call [`Pool::bind_tcp_listener`].
 ///
-/// [`Catalog`]: struct.Catalog.html
-/// [`Catalog::bind_tcp_listener`]: struct.Catalog.html#method.bind_tcp_listener
+/// [`Pool`]: struct.Pool.html
+/// [`Pool::bind_tcp_listener`]: struct.Pool.html#method.bind_tcp_listener
 pub struct TcpListener {
     std: net::TcpListener,
 }
